+++ modules/apps/sync/sync-service/src/main/java/com/liferay/sync/model/listener/ResourcePermissionModelListener.java
@@ -14,31 +14,22 @@
 
 package com.liferay.sync.model.listener;
 
-import com.liferay.document.library.kernel.model.DLFileEntry;
-import com.liferay.document.library.kernel.model.DLFolder;
 import com.liferay.portal.kernel.exception.ModelListenerException;
-import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.ModelListener;
 import com.liferay.portal.kernel.model.ResourcePermission;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.service.ResourcePermissionLocalService;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.sync.constants.SyncDLObjectConstants;
 import com.liferay.sync.model.SyncDLObject;
-import com.liferay.sync.service.SyncDLObjectLocalService;
 
 import java.util.Date;
-import java.util.List;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Shinn Lok
  */
 @Component(immediate = true, service = ModelListener.class)
 public class ResourcePermissionModelListener
-	extends BaseModelListener<ResourcePermission> {
+	extends SyncBaseModelListener<ResourcePermission> {
 
 	@Override
 	public void onBeforeCreate(ResourcePermission resourcePermission)
@@ -71,7 +62,7 @@ public class ResourcePermissionModelListener
 			syncDLObject.setModifiedTime(date.getTime());
 			syncDLObject.setLastPermissionChangeDate(date);
 
-			_syncDLObjectLocalService.updateSyncDLObject(syncDLObject);
+			syncDLObjectLocalService.updateSyncDLObject(syncDLObject);
 		}
 	}
 
@@ -86,7 +77,7 @@ public class ResourcePermissionModelListener
 		}
 
 		ResourcePermission originalResourcePermission =
-			_resourcePermissionLocalService.fetchResourcePermission(
+			resourcePermissionLocalService.fetchResourcePermission(
 				resourcePermission.getResourcePermissionId());
 
 		if (originalResourcePermission.hasActionId(ActionKeys.VIEW) &&
@@ -97,7 +88,7 @@ public class ResourcePermissionModelListener
 			syncDLObject.setModifiedTime(date.getTime());
 			syncDLObject.setLastPermissionChangeDate(date);
 
-			_syncDLObjectLocalService.updateSyncDLObject(syncDLObject);
+			syncDLObjectLocalService.updateSyncDLObject(syncDLObject);
 		}
 		else if (!originalResourcePermission.hasActionId(ActionKeys.VIEW) &&
 				 resourcePermission.hasActionId(ActionKeys.VIEW)) {
@@ -105,61 +96,5 @@ public class ResourcePermissionModelListener
 			updateSyncDLObject(syncDLObject);
 		}
 	}
-
-	protected SyncDLObject getSyncDLObject(
-		ResourcePermission resourcePermission) {
-
-		String modelName = resourcePermission.getName();
-
-		if (modelName.equals(DLFileEntry.class.getName())) {
-			return _syncDLObjectLocalService.fetchSyncDLObject(
-				SyncDLObjectConstants.TYPE_FILE,
-				GetterUtil.getLong(resourcePermission.getPrimKey()));
-		}
-		else if (modelName.equals(DLFolder.class.getName())) {
-			return _syncDLObjectLocalService.fetchSyncDLObject(
-				SyncDLObjectConstants.TYPE_FOLDER,
-				GetterUtil.getLong(resourcePermission.getPrimKey()));
-		}
-
-		return null;
-	}
-
-	@Reference(unbind = "-")
-	protected void setResourcePermissionLocalService(
-		ResourcePermissionLocalService resourcePermissionLocalService) {
-
-		_resourcePermissionLocalService = resourcePermissionLocalService;
-	}
-
-	@Reference(unbind = "-")
-	protected void setSyncDLObjectLocalService(
-		SyncDLObjectLocalService syncDLObjectLocalService) {
-
-		_syncDLObjectLocalService = syncDLObjectLocalService;
-	}
-
-	protected void updateSyncDLObject(SyncDLObject syncDLObject) {
-		syncDLObject.setModifiedTime(System.currentTimeMillis());
-
-		_syncDLObjectLocalService.updateSyncDLObject(syncDLObject);
-
-		String type = syncDLObject.getType();
-
-		if (!type.equals(SyncDLObjectConstants.TYPE_FOLDER)) {
-			return;
-		}
-
-		List<SyncDLObject> childSyncDLObjects =
-			_syncDLObjectLocalService.getSyncDLObjects(
-				syncDLObject.getRepositoryId(), syncDLObject.getTypePK());
-
-		for (SyncDLObject childSyncDLObject : childSyncDLObjects) {
-			updateSyncDLObject(childSyncDLObject);
-		}
-	}
-
-	private ResourcePermissionLocalService _resourcePermissionLocalService;
-	private SyncDLObjectLocalService _syncDLObjectLocalService;
 
 }
